% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBamLoader.R
\name{ggBamLoader}
\alias{ggBamLoader}
\title{Load a BAM/SAM file and convert it to a DataFrame.}
\usage{
ggBamLoader(bampath)
}
\arguments{
\item{bampath}{A string representing the path to the input BAM or SAM file.}
}
\value{
An S4 DataFrame containing the mapped reads from the file.
}
\description{
This function loads a BAM/SAM file, checks if the format is correct, and
converts it to an S4 class DataFrame for easier manipulation.
}
\examples{
# Load a BAM file from the package's extdata folder
df_gg <- ggBamLoader(system.file("extdata", "subset.bam", package = "GGCigarAligner"))
}
