% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCigarAligner.R
\name{ggCigarAligner}
\alias{ggCigarAligner}
\title{Align a Read to a Reference Genome}
\usage{
ggCigarAligner(df_gg, qname = NULL, index = NULL, my_reference)
}
\arguments{
\item{df_gg}{A S4 class DataFrame containing BAM/SAM alignment data, as obtained from
the \code{ggbamloader} function.}

\item{qname}{A character string specifying the query name (optional). If
provided, the function will attempt to select the read(s) with this
\code{qname}. If multiple reads have the same \code{qname}, the \code{index} parameter
should be used instead.}

\item{index}{An integer specifying the row index in the DataFrame (optional).
If provided, this will be used to select the specific read regardless of
the \code{qname}. This parameter takes precedence over \code{qname} if both are
provided.}

\item{my_reference}{A character string specifying the name of the reference
genome to be used (e.g., "BSgenome.Hsapiens.UCSC.hg38"). The genome must be
available in the \code{BSgenome} package.}
}
\value{
A list containing:
\item{reference_sequence}{A \code{DNAString} object representing the aligned reference sequence.}
\item{aligned_sequence}{A \code{DNAString} object representing the aligned read sequence.}
\item{original_sequence}{The original sequence from the read as a character string.}
}
\description{
This function aligns a specified read to its reference genome based on its
CIGAR string.
The user can select a read either by its query name (\code{qname}) or by its index
in the DataFrame. If multiple reads have the same \code{qname}, it is recommended
to use the index parameter to specify the exact read to be processed.
}
\details{
If the \code{qname} parameter is specified, the function will search for reads with this \code{qname}. If more than
one read is found, the function will notify the user and return \code{NULL} if the \code{index} parameter is not used.
If the \code{index} parameter is used, it will directly select the read at the given index. The \code{qname}
parameter will be ignored in this case. The \code{index} parameter must be a valid row index in the DataFrame.
}
\examples{
# Create a small mock df_gg DataFrame
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly = TRUE)) {
  df_gg <- ggBamLoader(system.file("extdata", "subset.bam", package = "GGCigarAligner"))
  
  # Align read using qname
  result1 <- ggCigarAligner(df_gg, qname = "ERR188273.4711308", my_reference = "BSgenome.Hsapiens.UCSC.hg38")
  
  # Align read using index
  result2 <- ggCigarAligner(df_gg, index = 3, my_reference = "BSgenome.Hsapiens.UCSC.hg38")
  
} else {
  message("Skipping example as BSgenome.Hsapiens.UCSC.hg38 is not installed.")
}
}
